♻️ Refactor: seperate character card
캐릭터 카드 분리
